Added yesterdays methods to Driver
I added the functions I wrote yesterday to the driver and edited them a
bit so they worked with the implementation (still have not tested
recordStay). Also, I changed all static functions to include a scanner
argument, since having multiple scanners (one in the main method and one
in the current method being run) was causing problems. So now the
scanner in the main method is passed into the functions.