Added functions 10, 12, 13, and 14
The functions for numbers 10, 12, 13, and 14 are now implemented. Not
every aspect could be tested since there weren't entries in Reserves and
a few other tables, but the syntax should be good.
Also, for recordStay (and reserve), the main method needs to store the
stays and reservations the user wishes to record, then prompt them on
exiting if they want to confirm those records. I didn't implement this,
but I wanted to make a note that when a reservation is made, an entry
needs to be added to the Period Table as well to save the begining and
end date of the reservation.